Add getSlotClassNameProp to allow custom style hooks to preserve the original className while overriding the component default className. (#34166)
Save each slot's user-provided class name (the `className` prop) in a metadata field on the slot object. This allows it to be accessed later, even after the default styles are applied. Provide a new function `getSlotClassNameProp_unstable` to access this value.

Each custom style hook will need to be updated to append a call to `getSlotClassNameProp_unstable` at the end of each call to `mergeClasses`.